# public/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Classic Mart External API",
    "version": "1.0.0",
    "description": "Versioned mobile and seller integration APIs backed by the same Classic Mart catalogue, cart, order, payment, inventory and fulfilment authority as the web application."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Classic Mart origin"
    }
  ],
  "paths": {
...
    "/api/v1/mobile/products/{id}": {
      "get": {
        "summary": "Read one published product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published product"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/v1/seller/products": {
      "get": {
        "summary": "List products for the API key store",
        "security": [
          {
            "sellerApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped products"
          }
        }
      }
    },
...
    "/api/v1/seller/orders": {
      "get": {
        "summary": "List seller orders for the API key store",
        "security": [
          {
            "sellerApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped seller orders"
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor returned by page.next."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "processing",
                "shipped",
                "delivered",
                "pending_payment",
                "cancelled"
              ],
              "default": "all"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ]
      }
    },
...
    "/api/v1/seller/orders/{id}": {
      "get": {
        "summary": "Read the scoped seller order, delivery contact and reservation warehouses",
        "security": [
          {
            "sellerApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Seller order with current version and permitted fulfilment actions. Internal stock costs and carrier proof codes are excluded."
          },
          "404": {
            "description": "Seller order unavailable in this store."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "mobileBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Opaque Classic Mart mobile access token"
      },
      "sellerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Classic Mart seller API key"
      }
    },
...
  },
...
}